Workerize route generation and add async updates
