+++ package.json
@@ -76,6 +76,7 @@
     "gulp-watch": "^4.3.11",
     "nodemon": "^1.11.0",
     "react-hot-loader": "^1.3.0",
-    "webpack-dev-server": "^2.2.1"
+    "webpack-dev-server": "^2.2.1",
+    "worker-loader": "^0.8.1"
   }
 }
